requirements.txt
cryptography==39.0.2
dateparser==1.1.7
fastapi==0.93.0
loglifos==0.1.0
motor==3.1.1
passlib==1.7.4
pyjwt==2.6.0
python-creditcard==0.0.1
python-decouple==3.8
python-multipart==0.0.6
uvicorn==0.20.0
uvloop==0.17.0
